requirements.txt
pygame
hurry.filesize
